# src/backend/internal/server/v2/docs/openapi.yaml
openapi: '3.0.2'
info:
  title: Спортивные залы
  version: v2
...
paths:
...
  /api/v2/client/{clientId}/schedule:
    get:
      operationId: getListSchedules
      tags:
        - Schedule
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid  
      responses:
        '200':
          description: "Записи на тренировку получены"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Записи на тренировку не найдены"  
        '500':
          description: "Внутренняя ошибка сервера"
    post:
      operationId: postSchedule
      tags:
        - Schedule
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '201':
          description: "Запись на тренировку создана"
        '400':
          description: "Некорректный запрос"  
        '401':
          description: "Неавтризован"
        '500':
          description: "Внутренняя ошибка сервера"
  /api/v2/gym/{clientId}/schedule/{scheduleId}:        
    put:
      operationId: putSchedule
      tags:
        - Schedule
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchSchedule'
      responses:
        '204':
          description: "Запись на тренировку обновлена"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Записи на тренировку не найдены"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteSchedule
      tags:
        - Schedule
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: scheduleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Запись на тренировку удалена"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Запись на тренировку не найдена"
        '500':
          description: "Внутренняя ошибка сервера"          
...
  /api/v2/trainer:
      post:
        operationId: postTrainer
        tags:
          - Trainer
        requestBody:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainer'
        responses:
          '201':
            description: "Тренер создан"
          '400':
            description: "Некорректный запрос"  
          '401':
            description: "Неавтризован"
          '500':
            description: "Внутренняя ошибка сервера"
      get:
        operationId: getListTrainers
        tags:
          - Trainer
        responses:
          '200':
            description: "Тренеры получены"
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Trainer'
          '404':
            description: "Тренеры не найдены"  
          '500':
            description: "Внутренняя ошибка сервера"      
  /api/v2/trainer/{trainerId}:
    get:
      operationId: getTrainer
      tags:
        - Trainer
      parameters:
        - name: trainerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: "Тренер получен"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trainer'
        '400':
          description: "Некорректный запрос"  
        '404':
          description: "Тренер не найден"  
        '500':
          description: "Внутренняя ошибка сервера"
    put:
      operationId: putTrainer
      tags:
        - Trainer
      parameters:
        - name: trainerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutTrainer'
      responses:
        '204':
          description: "Тренер обновлен"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Тренер не найден"
        '500':
          description: "Внутренняя ошибка сервера"
    delete:
      operationId: deleteTrainer
      tags:
        - Trainer
      parameters:
        - name: trainerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: "Тренер удален"
        '400':
          description: "Некорректный запрос"
        '401':
          description: "Неавтризован"  
        '404':
          description: "Тренер не найден"
        '500':
          description: "Внутренняя ошибка сервера"        
...
components:
  schemas:
...
    Trainer:
      type: object
      required:
        - id
        - fullname
        - email
        - phone
        - qualification
        - unitprice
        - gymsid
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор тренера в формате UUID"
        fullname:
          type: string
          description: "Полное имя тренера"
        email:
          type: string
          format: email
          description: "Электронная почта тренера"
        phone:
          type: string
          description: "Телефон тренера"
        qualification:
          type: string
          description: "Квалификация тренера"
        unitprice:
          type: number
          format: float
          description: "Цена за единицу услуг тренера"
        gymsid:
          type: array
          items:
            type: string
            format: uuid
          description: "Список уникальных идентификаторов спортзалов, с которыми связан тренер в формате UUID"
    PutTrainer:
      type: object
      required:
        - fullname
        - email
        - phone
        - qualification
        - unitprice
        - gymsid
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор тренера в формате UUID"
        fullname:
          type: string
          description: "Полное имя тренера"
        email:
          type: string
          format: email
          description: "Электронная почта тренера"
        phone:
          type: string
          description: "Телефон тренера"
        qualification:
          type: string
          description: "Квалификация тренера"
        unitprice:
          type: number
          format: float
          description: "Цена за единицу услуг тренера"
        gymsid:
          type: array
          items:
            type: string
            format: uuid
          description: "Список уникальных идентификаторов спортзалов, с которыми связан тренер в формате UUID"
...
    Schedule:
      type: object
      required:
        - id
        - dayoftheweek
        - starttime
        - endtime
        - clientid
        - trainingid
      properties:
        id:
          type: string
          format: uuid
          description: "Уникальный идентификатор расписания в формате UUID"
        dayoftheweek:
          type: string
          description: "День недели, когда запланирована тренировка"
        starttime:
          type: string
          format: time
          description: "Время начала тренировки в формате HH:MM"
        endtime:
          type: string
          format: time
          description: "Время окончания тренировки в формате HH:MM"
        clientid:
          type: string
          format: uuid
          description: "Уникальный идентификатор клиента в формате UUID"
        trainingid:
          type: string
          format: uuid
          description: "Уникальный идентификатор тренировки в формате UUID"
    PatchSchedule:
      type: object
      required:
        - id
        - dayoftheweek
        - starttime
        - endtime
        - clientid
        - trainingid
      properties:
        dayoftheweek:
          type: string
          description: "День недели, когда запланирована тренировка"
        starttime:
          type: string
          format: time
          description: "Время начала тренировки в формате HH:MM"
        endtime:
          type: string
          format: time
          description: "Время окончания тренировки в формате HH:MM"
  securitySchemes:
    Bearer:
      type: http
      description: Please enter a valid token
      scheme: bearer
      bearerFormat: JWT